# services/order-gateway/src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Order Gateway API",
    "version": "1.0.0",
    "description": "API documentation for the IUT BarakahBites Order Gateway"
  },
  "servers": [
    {
      "url": "http://localhost:3002",
      "description": "Local developer server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/items": {
      "get": {
        "summary": "List available items",
        "description": "Retrieves the list of menu items and their current stock levels.",
        "responses": {
          "200": {
            "description": "A list of items"
          }
        }
      }
    },
    "/order": {
      "post": {
        "summary": "Place an order",
        "description": "Places an order for one or more items. Includes sliding-window rate limiting (max 5 orders per 60s).",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "itemId": {
                          "type": "string",
                          "example": "iftar-box-1"
                        },
                        "quantity": {
                          "type": "integer",
                          "example": 1
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Order accepted"
          },
          "400": {
            "description": "Invalid format or missing items"
          },
          "401": {
            "description": "Unauthorized (Missing or invalid JWT)"
          },
          "409": {
            "description": "Insufficient stock"
          },
          "429": {
            "description": "Too many requests"
          }
        }
      }
    },
    "/order/{orderId}": {
      "get": {
        "summary": "Get order status",
        "description": "Retrieves the current cached state of an order.",
        "parameters": [
          {
            "in": "path",
            "name": "orderId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order details"
          },
          "404": {
            "description": "Order not found"
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Service Health",
        "description": "Returns the health status and dependencies.",
        "responses": {
          "200": {
            "description": "Healthy"
          }
        }
      }
    }
  }
}